Fixes 1122707 - Move the BrowserLocationView to its own file
